pyproject.toml
[project]
name = "analyze-bot"
version = "0.7.0"
description = "This is a Discord bot that analyzes the frequency of words used in messages and generates a word cloud."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect>=0.8.18",
    "discord-py>=2.7.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "pymongo>=4.16.0",
    "scipy>=1.17.1",
    "sudachidict-core>=20260116",
    "sudachipy>=0.6.10",
    "wordcloud>=1.9.6",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.2",
    "pytest>=9.0.2",
]
